change cart item quantity to double

+++ database/migrations/2021_02_05_000002_create_cart_items_table.php
@@ -17,9 +17,7 @@ class CreateCartItemsTable extends Migration {
             $table->bigIncrements('id');
             $table->foreignId('cart_id')->constrained()->onDelete("cascade");
             $table->foreignId('parent_id')->nullable()->constrained('cart_items')->onDelete("cascade");
-//            $table->integer('cart_id')->unsigned();
-//            $table->integer('parent_id')->unsigned()->nullable();
-            $table->decimal('quantity')->default(0);
+            $table->double('quantity')->default(0);
             $table->string('plu');
             $table->integer('type')->default(1);
             $table->decimal('weight', 12, 4)->default(1);
